# backend/ms-geolocalisation/src/main/java/com/pgvpt/controller/ZoneGeographiqueController.java
package com.pgvpt.controller;

import com.pgvpt.api.*;
import com.pgvpt.dto.*;
import com.pgvpt.service.ZoneService;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

import java.util.UUID;


@RestController
@RequiredArgsConstructor
public class ZoneGeographiqueController implements ZoneGeographiqueApi {

    private final ZoneService service;

    @Override
    public ResponseEntity<ZoneGeographique> createZoneGeographique(ZoneGeographiqueCreate zoneGeographiqueCreate) {
        return ResponseEntity.status(HttpStatus.CREATED).body(service.createGeo(zoneGeographiqueCreate));
    }

    @Override
    public ResponseEntity<Void> deleteZoneGeographique(UUID id) {
        service.deleteGeo(id);
        return ResponseEntity.noContent().build();
    }

    @Override
    public ResponseEntity<ZoneGeographique> getZoneGeographique(UUID id) {
        return ResponseEntity.ok(service.getByIdGeo(id));
    }

    @Override
    public ResponseEntity<PageZoneGeographique> getZonesGeographiques(Integer page, Integer size) {
        Pageable pageable = PageRequest.of(page, size);
        return ResponseEntity.ok(service.getAllGeo(pageable));
    }

    @Override
    public ResponseEntity<ZoneGeographique> updateZoneGeographique(UUID id, ZoneGeographiqueUpdate zoneGeographiqueUpdate) {
        return ResponseEntity.ok(service.updateGeo(id, zoneGeographiqueUpdate));
    }
}