# backend/ms-geolocalisation/src/main/resources/openapi/api.yml
openapi: 3.0.3
info:
  title: PGVPT - MS Géolocalisation API
  description: Propriétaire des coordonnées, du référentiel administratif, des zones géographiques et touristiques, des points
    d'accès et des recherches spatiales. Une géolocalisation cible exactement un patrimoine ou une entreprise; les références
    sont des UUID.
  version: 14.0.0
  contact:
    name: PGVPT
    email: contact@pgvpt.sn
servers:
  - url: http://localhost:3002/api/v1/geolocalisations
    description: Environnement local
  - url: https://api.pgvpt.sn/geolocalisation/api/v1
    description: Environnement de production
...
paths:
...
  /zones-geographiques:
    get:
      tags:
        - Zone geographique
      summary: Lister les Zone geographique
      operationId: getZonesGeographiques
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste des Zone geographique
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageZoneGeographique'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Zone geographique
      summary: Créer une zone geographique
      operationId: createZoneGeographique
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneGeographiqueCreate'
      responses:
        '201':
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneGeographique'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      parameters: []
      security:
        - bearerAuth: []
  /zones-geographiques/{id}:
    get:
      tags:
        - Zone geographique
      summary: Obtenir une zone géographique
      operationId: getZoneGeographique
      responses:
        '200':
          description: Zone trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneGeographique'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Zone geographique
      summary: Modifier un zone geographique
      operationId: updateZoneGeographique
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneGeographiqueUpdate'
      responses:
        '200':
          description: Zone geographique modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneGeographique'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Zone geographique
      summary: Supprimer une zone geographique
      operationId: deleteZoneGeographique
      responses:
        '204':
          description: Zone geographique supprimée
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/ZoneGeographiqueId'
...
components:
  parameters:
...
    ZoneGeographiqueId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      description: Numéro de page, indexé à partir de 0.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Nombre maximal d’éléments par page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Unauthorized:
      description: Authentification requise
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Accès interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    BaseEntity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      description: Classe racine commune de toutes les entités persistantes du PGVPT.
...
    ZoneGeographique:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      required:
        - id
        - code
        - nom
        - regions
      properties:
        code:
          type: string
          example: ZG-OUEST
        nom:
          type: string
          example: Zone géographique Ouest
        description:
          type: string
        regions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ZoneGeographiqueRegion'
        superficieKm2:
          type: number
          format: double
          minimum: 0
        latitudeCentre:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitudeCentre:
          type: number
          format: double
          minimum: -180
          maximum: 180
        actif:
          type: boolean
          default: true
      description: Zone géographique composée d’une ou plusieurs régions via ZoneGeographiqueRegion. Une région ne peut appartenir
        qu’à une seule zone géographique.
    ZoneGeographiqueCreate:
      type: object
      required:
        - code
        - nom
        - regionIds
      properties:
        code:
          type: string
          example: ZG-OUEST
        nom:
          type: string
          example: Zone géographique Ouest
        description:
          type: string
        regionIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
        superficieKm2:
          type: number
          format: double
          minimum: 0
        latitudeCentre:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitudeCentre:
          type: number
          format: double
          minimum: -180
          maximum: 180
    ZoneGeographiqueUpdate:
      type: object
      properties:
        code:
          type: string
          example: ZG-OUEST
        nom:
          type: string
          example: Zone géographique Ouest
        description:
          type: string
        regionIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        superficieKm2:
          type: number
          format: double
          minimum: 0
        latitudeCentre:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitudeCentre:
          type: number
          format: double
          minimum: -180
          maximum: 180
...
    ApiError:
      type: object
      required:
        - timestamp
        - status
        - code
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: object
          additionalProperties: true
...
    ZoneGeographiqueRegion:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      required:
        - id
        - zoneGeographiqueId
        - regionId
      properties:
        zoneGeographiqueId:
          type: string
          format: uuid
        regionId:
          type: string
          format: uuid
      description: Relation persistée entre une zone géographique et une région. Une région ne peut appartenir qu’à une seule
        zone géographique. Cette relation hérite de BaseEntity pour assurer la traçabilité (createdAt, updatedAt).
      x-constraints:
        - UNIQUE(regionId)
        - UNIQUE(zoneGeographiqueId, regionId)
      x-database-constraints:
        unique:
          - regionId
      x-business-rules:
        - Une région appartient à une seule zone géographique.
...
    PageZoneGeographique:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ZoneGeographique'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []